Move mixin definitions to TextAnnotation
